Move RollChannel to memoized component
